Added tkltest_unit module, updated test cases, extracted common code to be reused in the future tkltest_ui module in tkltest module (#151)
